- Save publications without image - Show publications saved without image

+++ app/src/main/java/com/svalero/toteco_app/domain/Establishment.java
@@ -12,15 +12,15 @@ public class Establishment {
     @ColumnInfo
     private String name;
     @ColumnInfo
-    private float latitude;
+    private double latitude;
     @ColumnInfo
-    private float longitude;
+    private double longitude;
     @ColumnInfo
     private boolean open;
     @ColumnInfo
     private float punctuation;
 
-    public Establishment(String name, float latitude, float longitude, boolean open, float punctuation) {
+    public Establishment(String name, double latitude, double longitude, boolean open, float punctuation) {
         this.name = name;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -44,19 +44,19 @@ public class Establishment {
         this.name = name;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
